requirements.txt
# Core ML libraries
torch>=1.12.0
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.20.0

# Detoxify dependencies
detoxify>=0.5.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Text processing
nltk>=3.7
spacy>=3.4.0
langdetect>=1.0.9
textblob>=0.17.1

# Configuration and utilities
pyyaml>=6.0
argparse
pathlib
tqdm>=4.64.0
joblib>=1.1.0

# Evaluation and metrics
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Logging and monitoring
tensorboard>=2.9.0
wandb>=0.13.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Jupyter and notebooks
jupyter>=1.0.0
ipykernel>=6.15.0
notebook>=6.4.0

# Data validation
great-expectations>=0.15.0

# Statistical analysis
statsmodels>=0.13.0

# Progress bars
rich>=12.0.0

# File handling
openpyxl>=3.0.0
xlrd>=2.0.0

# API and web
requests>=2.28.0
flask>=2.2.0

# Memory optimization
psutil>=5.9.0

# GPU support (optional)
# torch-audio>=0.12.0
# torch-vision>=0.13.0

# Multilingual support
polyglot>=16.7.4
# Note: polyglot requires additional system dependencies
# See: https://polyglot.readthedocs.io/en/latest/Installation.html

# Translation services (optional)
googletrans>=4.0.0
# Note: Google Translate API may require API key

# Additional NLP libraries
sentence-transformers>=2.2.0
wordcloud>=1.8.0

# Configuration management
hydra-core>=1.2.0
omegaconf>=2.2.0

# Model serving (optional)
fastapi>=0.78.0
uvicorn>=0.18.0

# Data versioning (optional)
dvc>=2.8.0

# Monitoring and alerting
mlflow>=1.28.0

# Performance profiling
py-spy>=0.3.0
memory-profiler>=0.60.0